# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cricket Knowledge RAG Chatbot API",
    "version": "1.0.0",
    "description": "API documentation for the Cricket Knowledge RAG (Retrieval-Augmented Generation) Chatbot. This backend parses documents (PDF, DOCX, TXT), chunks them, embeds them with the Gemini API, stores them in MongoDB Atlas, performs Vector Search, manages phone authentication, and provides personalized quizzes with score points."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user directly",
        "description": "Registers a new phone number and name. No OTP required to register. Fails if user is already registered.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890",
                    "description": "User phone number in E.164 format."
                  },
                  "name": {
                    "type": "string",
                    "example": "Cricket Fan",
                    "description": "User full name."
                  }
                },
                "required": ["phoneNumber", "name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "User registered successfully. You can now login." },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "phoneNumber": { "type": "string" },
                        "name": { "type": "string" },
                        "cricPoints": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. user already registered or missing parameters)."
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Request login OTP for phone number",
        "description": "Verifies that the phone number is registered, generates a mock OTP, logs it to console, and stores it for verification. Fails if user does not exist.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  }
                },
                "required": ["phoneNumber"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP generated successfully and logged to console.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP generated and logged to backend console successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User is not registered, or invalid phone number."
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP and complete login",
        "description": "Verifies the 6-digit OTP generated for the phone number. Returns JWT token for authorization.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "otp": {
                    "type": "string",
                    "example": "123456"
                  }
                },
                "required": ["phoneNumber", "otp"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. JWT token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5..."
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "phoneNumber": { "type": "string" },
                        "name": { "type": "string" },
                        "cricPoints": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP."
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "summary": "Get authenticated user profile and quiz history",
        "description": "Retrieves the user's information, cumulative Cric Points, and history of previous quiz attempts.",
        "responses": {
          "200": {
            "description": "Profile data retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "phoneNumber": { "type": "string" },
                        "name": { "type": "string" },
                        "cricPoints": { "type": "integer" },
                        "createdAt": { "type": "string", "format": "date-time" }
                      }
                    },
                    "attempts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "attemptId": { "type": "string" },
                          "quizId": { "type": "string" },
                          "score": { "type": "integer" },
                          "pointsEarned": { "type": "integer" },
                          "completedAt": { "type": "string", "format": "date-time" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized token."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}